% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_gcp.R
\name{choose_gcp}
\alias{choose_gcp}
\title{Function to choose the Ground Control Point (GCP) on an image}
\usage{
choose_gcp(image_obj, gcp = NULL, col = "red")
}
\arguments{
\item{image_obj}{An array representing the image (colour images are generally
imported as an array of nx x ny x 3 colour channels), or a file path to the
image (currenly this can only be of type .jpg).}

\item{gcp}{(optional) A dataframe of GCPs, containing columns \code{id}, \code{x},
\code{y},\code{longitude}, \code{latitude}. This is used if we want to add additional GCP
to an existing list (usually created by running this function multiple
times).}

\item{col}{The colour of the points to be plotted on the image. Default is
"red".}
}
\value{
A dataframe with the GCPs, including the image coordinates and their
corresponding geographic coordinates.
}
\description{
GCPs are used to georeference images by providing known coordinates for
specific points in the image. This functions extracts the pixel coordinates
(coded as the origin on the bottom left corner, see below) of the GCPs from
the image, and allows the user to input their corresponding geographic
coordinates (longitude and latitude).
}
\details{
NOTE: There are two conventions on how to define pixel coordinates. In this
function (and more generally throughout \code{crstools}), the origin is defined as
the bottom left corner of the image, with x increasing to the right and y
increasing upwards. This is consistent with the convention used in many image
processing libraries. However, some libraries (like OpenCV and GDAL) define
the origin at the top left corner, with y increasing downwards. Be sure to
check which convention your image processing library uses when working with
pixel coordinates.
}
\examples{
\dontshow{if (rlang::is_interactive() ) withAutoprint(\{ # examplesIf}
  # Get the path to an example image included in the package and choose GCPs
  img_path <- system.file("extdata/europe_map.jpeg", package = "crstools") 
  # this will open a new window where you can choose some points
  gcp_europe <- choose_gcp(img_path)
  # after the first set of points is chosen, we can add more points 
  gcp_europe <- choose_gcp(img_path, gcp = gcp_europe)
\dontshow{\}) # examplesIf}
}
