% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_ew_extent.R
\name{crs_ew_extent}
\alias{crs_ew_extent}
\title{East-West extent}
\usage{
crs_ew_extent(
  distortion,
  center,
  lonmin,
  lonmax,
  latmin,
  latmax,
  quiet = FALSE
)
}
\arguments{
\item{distortion}{character string (e.g., "equal_area", "conformal").}

\item{center}{numeric value, center of the map projection.}

\item{lonmin}{numeric value, minimum longitude of the map.}

\item{lonmax}{numeric value, maximum longitude of the map.}

\item{latmin}{numeric value, minimum latitude of the map.}

\item{latmax}{numeric value, maximum latitude of the map.}

\item{quiet}{logical, whether to suppress messages.}
}
\value{
data.frame with the suggested projection.
}
\description{
Checking the East-West extent of a regional map
}
\keyword{internal}
