% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_hemisphere.R
\name{crs_hemisphere}
\alias{crs_hemisphere}
\title{Hemisphere projection}
\usage{
crs_hemisphere(distortion, center, scale, latmin, latmax)
}
\arguments{
\item{distortion}{character string (e.g., "Equalarea", "Equidistant", "Compromise").}

\item{center}{data.frame with two numeric values, latitude and longitude of the center of the map.}

\item{scale}{numeric value, scale of the map.}

\item{latmin}{numeric value, minimum latitude of the map.}

\item{latmax}{numeric value, maximum latitude of the map.}
}
\value{
data.frame with the suggested projection.
}
\description{
Function to suggest a projection for maps showing a hemisphere
}
\keyword{internal}
