% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_ns_extent.R
\name{crs_ns_extent}
\alias{crs_ns_extent}
\title{North-South extent}
\usage{
crs_ns_extent(distortion, center, quiet = FALSE)
}
\arguments{
\item{distortion}{character string (e.g., "equal_area", "equidistant",
"conformal").}

\item{center}{The center of the map projection.}

\item{quiet}{logical, whether to suppress messages.}
}
\value{
data.frame with the suggested projection.
}
\description{
Checking the North-South extent of a regional map
}
\keyword{internal}
