% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_square_format.R
\name{crs_square_format}
\alias{crs_square_format}
\title{projections for a small map with square format}
\usage{
crs_square_format(property, center, latmin, latmax)
}
\description{
projections for a small map with square format
}
