% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_string.R
\name{crs_string}
\alias{crs_string}
\title{Function to format the PROJ.4 and WKT strings}
\usage{
crs_string(
  prj = c("aeqd", "laea", "stere", "aea", "eqdc", "lcc", "cea", "merc", "eqc", "eqearth",
    "moll", "hammer", "eck4", "wag4", "wag7", "tpeqd", "robin", "natearth", "wintri",
    "patterson", "latlong", "mill", "tmerc", "cass", "tcea"),
  x0,
  lat0,
  lat1,
  lat2,
  lon0,
  k0,
  datum = c("WGS84", "ETRS89", "NAD83"),
  unit = c("m", "ft")
)
}
\arguments{
\item{prj}{Character string. The projection type. Options are "aeqd"
(Azimuthal Equidistant), "laea" (Lambert Azimuthal Equal Area), "stere"
(Stereographic), "aea" (Albers Equal Area), "eqdc" (Equidistant Conic),
"lcc" (Lambert Conformal Conic), "cea" (Cylindrical Equal Area), "merc"
(Mercator), "eqc" (Equidistant Cylindrical), "eqearth" (Equal Earth),
"moll" (Mollweide), "hammer" (Hammer),
"eck4" (Eckert IV), "wag4" (Wagner IV),
"wag7" (Wagner VII), "tpeqd" (Two Point Equidistant), "robin" (Robinson),
"natearth" (Natural Earth), "wintri" (Winkel Tripel),
"patterson" (Patterson), "latlong" (Latitude/Longitude),
"mill" (Miller Cylindrical), "tmerc" (Transverse
Mercator), "cass" (Cassini), "tcea" (Transverse Cylindrical Equal Area).
.}

\item{x0}{Numeric. The false easting value.}

\item{lat0}{Numeric. The latitude of origin.}

\item{lat1}{Numeric. The first standard parallel.}

\item{lat2}{Numeric. The second standard parallel.}

\item{lon0}{Numeric. The central meridian.}

\item{k0}{Numeric. The scale factor at the central meridian.}

\item{datum}{Character string. The datum. Options are "WGS84", "ETRS89",
"NAD83".}

\item{unit}{Character string. The linear unit. Options are "m" (meters) and
"ft" (feet).}
}
\value{
A list with two elements: "proj4" and "wkt". The "proj4" element
contains the PROJ.4 string, while the "wkt" element contains the WKT
string.
}
\description{
This function takes the parameters of a custom projection and returns the
PROJ.4 and WKT strings.
}
\details{
Note that the WKT string contains a few special characters to define
indentation; these should not impact the parsing of the wkt string (eg. by
\code{sf})
}
\keyword{internal}
