% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_string.R
\name{crs_string_row}
\alias{crs_string_row}
\title{function to feed the row of a crs_df to crs_string}
\usage{
crs_string_row(x, datum, unit)
}
\arguments{
\item{x}{a vector from a row of a crs_df}

\item{datum}{Character string. The datum. Options are "WGS84", "ETRS89", "NAD83".}

\item{unit}{Character string. The linear unit. Options are "m" (meters) and "ft" (feet).}
}
\value{
A list with two elements: "proj4" and "wkt". The "proj4" element
contains the PROJ.4 string, while the "wkt" element contains the WKT string.
}
\description{
function to feed the row of a crs_df to crs_string
}
