% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_world_equidistant.R
\name{crs_world_equidistant}
\alias{crs_world_equidistant}
\title{Function to create the world equidistant projection}
\usage{
crs_world_equidistant(center, scale, round_cm = FALSE, projection)
}
\arguments{
\item{center}{The center of the map (named vector with lat and long)}

\item{scale}{The scale of the map}

\item{round_cm}{The round central meridian.}

\item{projection}{The projection to use
projection suggestions}
}
\value{
The projection object
}
\description{
Function to create the world equidistant projection
}
