% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_gcp.R
\name{choose_gcp}
\alias{choose_gcp}
\title{Function to choose the Ground Control Point (GCP) on an image}
\usage{
choose_gcp(image_obj, gcp = NULL, col = "red")
}
\arguments{
\item{image_obj}{An array representing the image (colour images are generally
imported as an array of nx x ny x 3 colour channels), or a file path to the
image (of type .jpg or XXXX).}

\item{gcp}{(optional) A dataframe of GCPs, containing columns \code{id}, \code{x},
\code{y},\code{longitude}, \code{latitude}. This is used if we want to add additional GCP
to an existing list (usually created by running this function multiple
times).}

\item{col}{The colour of the points to be plotted on the image. Default is
"red".}
}
\value{
A dataframe with the GCPs, including the image coordinates and their
corresponding geographic coordinates.
}
\description{
GCPs are used to georeference images by providing known coordinates for
specific points in the image.
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get the path to an example image included in the package and choose GCPs
img_path <- system.file("extdata/europe_map.jpeg", package = "crstools")
# choose some points
gcp_europe <- choose_gcp(img_path)
# after the first set of points is chosen, we can add more points
gcp_europe <- choose_gcp(img_path, gcp = gcp_europe)
\dontshow{\}) # examplesIf}
}
