% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_ew_extent.R
\name{crs_ew_extent}
\alias{crs_ew_extent}
\title{East-West extent}
\usage{
crs_ew_extent(
  distortion,
  property,
  center,
  scale,
  lonmin,
  lonmax,
  latmin,
  latmax
)
}
\arguments{
\item{distortion}{character string (e.g., "Equalarea", "Equidistant", "Compromise").}

\item{property}{The property of the world map projection (e.g., "Equalarea", "Equidistant", "Compromise").}

\item{center}{The center of the map projection.}

\item{scale}{numeric value, scale of the map.}

\item{lonmin}{The minimum longitude of the map. Default is -180.}

\item{lonmax}{The maximum longitude of the map. Default is 180.}

\item{latmin}{numeric value, minimum latitude of the map.}

\item{latmax}{numeric value, maximum latitude of the map.}
}
\value{
data.frame with the suggested projection.
}
\description{
Checking the East-West extent of a regional map
}
\keyword{internal}
