% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_small_area.R
\name{crs_small_area}
\alias{crs_small_area}
\title{choose the crs for a small area (region or country)}
\usage{
crs_small_area(
  distortion,
  center,
  scale,
  lonmin,
  lonmax,
  latmin,
  latmax,
  ang_unit = "degrees"
)
}
\description{
maps showing a smaller area
}
