% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_wizard.R
\name{crs_wizard}
\alias{crs_wizard}
\title{Suggestions for map projections depending on the extent}
\usage{
crs_wizard(
  x,
  distortion = c("equal_area", "conformal", "equidistant", "compromise"),
  roundCM = FALSE,
  return_best = TRUE,
  datum = c("WGS84", "ETRS89", "NAD83"),
  unit = c("m", "ft"),
  lat_check = TRUE
)
}
\arguments{
\item{distortion}{The type of distortion to be minimized. Options are
"equal_area", "conformal", "equidistant" and "compromise". Default is
"equal_area".}

\item{roundCM}{Logical. If TRUE, the central meridian is rounded to the
nearest degree. Default is FALSE.}

\item{return_best}{Logical. If TRUE, only the best projection is returned, otherwise,
if there are multiple options, a list will be returned}

\item{lat_check}{Logical. If TRUE, the function will check if lat values are within range (-90,90).
Default is TRUE.}

\item{lonmin}{The minimum longitude of the map. Default is -180.}

\item{lonmax}{The maximum longitude of the map. Default is 180.}

\item{latmin}{The minimum latitude of the map. Default is -90.}

\item{latmax}{The maximum latitude of the map. Default is 90.}
}
\value{
Either a list of two strings (proj4 and WKT) for a single projection
(if either only one projection is available or return_best is TRUE),
or a list of lists, each containing two strings (proj4 and WKT) for a
single projection (if multiple projections are available and return_best is FALSE).
}
\description{
This function provides suggestions for map projections depending on the extent
of the map. For SDMs, an equal area projection is generally favoured, but
it is also possible to get suggestions for conformal, equidistant or
compromise projections. The algorithm is a reimplementation in R of
Projection Wizard (Bojan et al., 2016), version 2.1.
}
\details{
TODO give some additional explanation about distortions, trade-offs
(especially in the case of whole world maps)
}
\examples{
# Whole map
crs_wizard(c(-180,180,-90, 90))
# Northen Hemisphere
crs_wizard(lonmin = -180, lonmax = 180, latmin = 21, latmax = 70)
# Hemisphere showing the tropics
crs_wizard(lonmin = -180, lonmax = 180, latmin = -7, latmax = 21)
# Regional map for EW extent
crs_wizard(lonmin = -60, lonmax = 20, latmin = 40, latmax = 70)
}
\references{
Bojan, S, Bernhard, J, & Helen, J (2016) Projection Wizard - An
Online Map Projection Selection Tool. The Cartographic Journal 53(2):177-185
DOI: 10.1080/00087041.2015.1131938
}
