% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_wizard.R
\name{crs_wizard}
\alias{crs_wizard}
\title{Suggestions for map projections depending on the extent}
\usage{
crs_wizard(
  distortion = c("Equalarea", "Comformal", "Equidistant", "Compromise"),
  lonmin = -180,
  lonmax = 180,
  latmin = -90,
  latmax = 90,
  roundCM = FALSE
)
}
\arguments{
\item{distortion}{The type of distortion to be minimized. Options are
"Equalarea", "Conformal", "Equidistant" and "Compromise". Default is
"Equalarea".}

\item{lonmin}{The minimum longitude of the map. Default is -180.}

\item{lonmax}{The maximum longitude of the map. Default is 180.}

\item{latmin}{The minimum latitude of the map. Default is -90.}

\item{latmax}{The maximum latitude of the map. Default is 90.}

\item{roundCM}{Logical. If TRUE, the central meridian is rounded to the
nearest degree. Default is FALSE.}
}
\value{
A data frame with the suggested map projections.
}
\description{
This function provides suggestions for map projections depending on the extent
of the map. For SDMs, an equal area projection is generally favoured, but
it is also possible to get suggestions for conformal, equidistant or
compromise projections. The algorithm is a reimplementation in R of
Projection Wizard (Bojan et al., 2016), version 2.1.
}
\examples{
# Whole map
crs_wizard(lonmin = -180, lonmax = 180,latmin = -90, latmax = 90)
# Northen Hemisphere
crs_wizard(lonmin = -180, lonmax = 180, latmin = 21, latmax = 70)
# Hemisphere showing the tropics
crs_wizard(lonmin = -180, lonmax = 180, latmin = -7, latmax = 21)
# Regional map for EW extent
crs_wizard(lonmin = -60, lonmax = 20, latmin = 40, latmax = 70)
}
\references{
Bojan, S, Bernhard, J, & Helen, J (2016) Projection Wizard - An
Online Map Projection Selection Tool. The Cartographic Journal 53(2):177-185
DOI: 10.1080/00087041.2015.1131938
}
