% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_world_equidistant.R
\name{crs_world_equidistant}
\alias{crs_world_equidistant}
\title{Function to create the world equidistant projection}
\usage{
crs_world_equidistant(
  center,
  scale,
  round_cm = FALSE,
  prj_details,
  quiet = FALSE
)
}
\arguments{
\item{center}{The center of the map (named vector with lat and long)}

\item{scale}{The scale of the map}

\item{round_cm}{The round central meridian.}

\item{prj_details}{a list of options defining the desired projection}

\item{quiet}{Whether to suppress messages and warnings}
}
\value{
The projection object
}
\description{
Function to create the world equidistant projection
}
\keyword{internal}
