% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crstools-package.R
\docType{package}
\name{crstools-package}
\alias{crstools}
\alias{crstools-package}
\title{crstools: Tools to Work with Projections}
\description{
Choose the appropriate map projection for a given application, visualise the resulting distortion, and georeference data from unknown projections.
}
\author{
\strong{Maintainer}: Andrea Manica \email{am315@cam.ac.uk} (\href{https://orcid.org/0000-0003-1895-450X}{ORCID})

Authors:
\itemize{
  \item Andrea Pozzi
  \item Cecilia Padilla-Iglesias
  \item Michela Leonardi
  \item Aramish Fatima
  \item Ondrej Pelanek
  \item Nile Stephenson
  \item Margherita Colucci
}

}
\keyword{internal}
