% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coords.R
\name{extract_coords}
\alias{extract_coords}
\title{Extract coordinates}
\usage{
extract_coords(georef_image, coords_df = NULL, col = "red")
}
\arguments{
\item{georef_image}{A spatraster object representing the georeferenced image.}

\item{coords_df}{An optional dataframe containing the coordinates of points
previously extracted from the image. Default is NULL.}

\item{col}{The colour of the points to be plotted on the image. Default is
"red".}
}
\value{
A dataframe with ID and coordinates of the points extracted from the
image.
}
\description{
Extract coordinates from a georeferenced image.
}
