% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tissot.R
\name{geom_tissot}
\alias{geom_tissot}
\title{Add Tissot's indicatrix to a map}
\usage{
geom_tissot(
  mapping = ggplot2::aes(),
  data = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  centres = c(5, 5),
  radius = NULL,
  fill = "red",
  alpha = 0.7,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}}. If
specified and inherit.aes = TRUE (the default), it is combined with the
default mapping at the top level of the plot. You must supply mapping if
there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are two options: A
sf object or a SpatRaster object}

\item{na.rm}{If FALSE, the default, missing values are removed with a
warning. If TRUE, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends? NA,
the default, includes if any aesthetics are mapped. FALSE never includes,
and TRUE always includes. It can also be a named logical vector to finely
select the aesthetics to display.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than
combining with them. This is most useful for helper functions that define
both data and aesthetics and shouldn't inherit behaviour from the default
plot specification, e.g. \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{centres}{Either a list with elements "lng" and "lat" giving the
longitudes and latitudes of the grid of centres for the Tissot's
indicatrix, or a vector of length 2 with the number of rows and columns to
generate such a grid automatically. The latter is passed to \code{\link[=pretty]{pretty()}} to
get pretty breaks, and the exact number of Tissot's circles might differ
from the input numbers (see \code{\link[=pretty]{pretty()}} for details. It defaults to c(5,5).}

\item{radius}{The radius of the circles (see \code{\link[sf:st_buffer]{sf::st_buffer()}} for an
explanation of how units are created; we recommend that you explicitly
state your using with the \code{units::as_units()}, e.g \code{units::as_units(100, "km")}}

\item{fill}{The fill colour of the circles}

\item{alpha}{The transparency of the circles}

\item{...}{Other arguments passed on to \code{\link[ggplot2:geom_sf]{ggplot2::geom_sf()}}}
}
\value{
A ggplot2 object
}
\description{
This function adds Tissot's indicatrix to a map. Tissot's indicatrix is a
mathematical contrivance used in cartography to characterize local
distortions due to map projection.
}
\examples{
\dontshow{if (rlang::is_installed("rnaturalearth")) withAutoprint(\{ # examplesIf}
# load required packages
library(rnaturalearth)
library(sf)
library(ggplot2)
s_america_sf <- ne_countries(continent = "South America", returnclass = "sf")
s_am_equal_area <- suggest_crs(s_america_sf, distortion = "equal_area")
ggplot(data = s_america_sf) +
 geom_sf() +
 geom_tissot() +
 coord_sf(crs = s_am_equal_area$proj4) +
 theme_minimal()
\dontshow{\}) # examplesIf}
}
