% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georeference_img.R
\name{georeference_img}
\alias{georeference_img}
\title{Georeference an Image using Ground Control Points (GCPs)}
\usage{
georeference_img(
  image_obj,
  gcp,
  output_path = NULL,
  transform_method = c("auto", "poly_1", "poly_2", "poly_3", "tps", "auto")
)
}
\arguments{
\item{image_obj}{A character string specifying the file path to the input
image (JPEG format). The function reads this image and applies the GCPs for
georeferencing.}

\item{gcp}{A data frame containing the Ground Control Points (GCPs). This
dataframe can be produced with the \emph{draw_gcp_points} function. This data
frame should have the following columns:
\itemize{
\item \code{id}: An identifier for each GCP (numeric).
\item \code{x}: The x-coordinate of the GCP (in pixel space).
\item \code{y}: The y-coordinate of the GCP (in pixel space).
\item \code{lon}: The longitude of the GCP (georeferenced).
\item \code{lat}: The latitude of the GCP (georeferenced).
}}

\item{output_path}{A character string representing the file path to the input
image. (\verb{_warp.tif}) will be appended to it.}

\item{transform_method}{A character string specifying the transformation
method to be used for warping the image. Options are "poly_1" (first order
polynomial), "poly_2" (second order polynomial), "poly_3" (third order
polynomial), "tps" (thin plate spline), or "auto" (the default, allowing
GDAL to choose of a polynomial of the appropriate order based on the
number of available GCP. Polynomials are best for standard maps directly
captured from a publication (a first or second order polynomial is often
sufficient), tps allows for scanning artefacts, but it is badly affected
by any incorrect GCP.}
}
\value{
A character string representing the path to the newly created warped
TIFF image file (\verb{_warp.tif}). This file contains the georeferenced image.
}
\description{
This function warps a given image using a set of Ground Control Points (GCPs)
to create a georeferenced version of the image. The function uses GDAL tools
to first translate the image into a georeferenced TIFF format and then
applies a warp operation to reproject the image into a spatial reference
system (EPSG:4326). The warped image is saved as a new file with the suffix
\verb{_warp.tif}.
}
\examples{
\dontshow{if (rlang::is_interactive()) withAutoprint(\{ # examplesIf}
# get the path to an example image included in the package
img_path <- system.file("extdata/europe_map.jpeg",
  package = "crstools"
)
# load a set of GCPs (or we could create them using the choose_gcp()
# and find_gcp() functions)
gcp_df <- readRDS(system.file(
  "extdata/europe_gcp_georef.RDS",
  package = "crstools"
))
# Assuming you have a set of GCPs in gcp_df and an image file "image.jpg"
warped_img <- georeference_img(
  image_obj = img_path, gcp = gcp_df,
  output_path = tempfile(
    pattern = "georef_img_", tmpdir = tempdir(),
    fileext = ".tif"
  )
)
\dontshow{\}) # examplesIf}
}
