% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georeference_img.R
\name{georeference_img}
\alias{georeference_img}
\title{Georeference an Image using Ground Control Points (GCPs)}
\usage{
georeference_img(image_obj, gcp, output_path = NULL)
}
\arguments{
\item{image_obj}{A character string specifying the file path to the input
image (JPEG format). The function reads this image and applies the GCPs for
georeferencing.}

\item{gcp}{A data frame containing the Ground Control Points (GCPs). This
dataframe can be produced withthe \emph{draw_gcp_points} function. This data
frame should have the following columns:
\itemize{
\item \code{id}: An identifier for each GCP (numeric).
\item \code{x}: The x-coordinate of the GCP (in pixel space).
\item \code{y}: The y-coordinate of the GCP (in pixel space).
\item \code{lon}: The longitude of the GCP (georeferenced).
\item \code{lat}: The latitude of the GCP (georeferenced).
}}

\item{output_path}{A character string representing the file path to the input
image. (\verb{_warp.tif}) will be appended to it.}
}
\value{
A character string representing the path to the newly created warped
TIFF image file (\verb{_warp.tif}). This file contains the georeferenced image.
}
\description{
This function warps a given image using a set of Ground Control Points (GCPs)
to create a georeferenced version of the image. The function uses GDAL tools
to first translate the image into a georeferenced TIFF format and then
applies a warp operation to reproject the image into a spatial reference
system (EPSG:4326). The warped image is saved as a new file with the suffix
\verb{_warp.tif}.
}
\examples{
\dontrun{
# Assuming you have a set of GCPs in gcp_df and an image file "image.jpg"
warped_image <- georeference_image(image_path = "image.jpg", gcp_df = gcp_df)
}
}
