% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gcp.R
\name{plot_gcp}
\alias{plot_gcp}
\title{Function to plot the Ground Control Point (GCP) on an image}
\usage{
plot_gcp(image_obj, gcp = NULL, col = "red")
}
\arguments{
\item{image_obj}{An array representing the image (colour images are generally
imported as an array of nx x ny x 3 colour channels), or a file path to the
image (currenly this can only be of type .jpg).}

\item{gcp}{A dataframe of GCPs, containing at least columns \code{id}, \code{x},
\code{y}.}

\item{col}{The colour of the points to be plotted on the image. Default is
"red".}
}
\value{
A plot with the GCPs
}
\description{
GCPs are used to georeference images by providing known coordinates for
specific points in the image. This function plots GCPs onto the original map
}
\examples{
\dontshow{if (rlang::is_interactive()) withAutoprint(\{ # examplesIf}
# Get the path to an example image included in the package
img_path <- system.file("extdata/europe_map.jpeg", package = "crstools")
# load some GCPs
gcp_europe <- readRDS(system.file(
                                "extdata/europe_gcp_georef.RDS",
                                package = "crstools"))
# plot
plot_gcp(img_path, gcp = gcp_europe)
\dontshow{\}) # examplesIf}
}
